Remove indicator parameter
I had put an indicator parameter into mic originally, which
would allow you to skip the beep or indicator that Naomi is
ready for you to talk, or has accepted your input and is
processing it. That does not exist in the current version
of Naomi though, so was causing an error.