Fix an issue when checking for error
When Zork returns an error (like "I don't know that word" or "you
used that word in a way I don't understand) the string starts with
a space and the location is listed as the starting position of
"West of house". Fixed a check for the first character of the
description being a space to work even if the description string
is empty.